sarvam_hindi_agent/requirements.txt
# Pipecat Voice Agent with LiveKit transport
pipecat-ai[sarvam,azure,livekit]
python-dotenv
loguru

# Token server
fastapi
uvicorn
livekit-api
